requirements.txt
amqp==5.1.1
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
celery==5.2.7
certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cron-descriptor==1.2.35
cryptography==42.0.4
decorator==5.1.1
Django==4.0
django-celery-beat==2.5.0
django-constance==2.9.1
django-cors-headers==3.14.0
django-filter==22.1
django-phonenumber-field==7.0.2
django-picklefield==3.1
django-timezone-field==5.0
django-json-widget==2.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-localized-fields
executing==1.2.0
firebase-admin==6.2.0
gunicorn==20.1.0
idna==3.4
ipython
jedi==0.18.2
kombu==5.2.4
matplotlib-inline==0.1.6
pandas==2.2.2
parso==0.8.3
pexpect==4.8.0
phonenumbers==8.13.6
pickleshare==0.7.5
prompt-toolkit
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyotp==2.8.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
redis==4.5.4
requests
simplejwt==2.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing==3.7.4.3
typing_extensions
tzdata==2023.3
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Pillow==10.0.1
django_redis==5.3.0
django-admin-sortable2==1.0.4
openpyxl==3.1.5
python-telegram-bot==12.6
asyncpg
aiogram
sqlalchemy>=2.0
geopy
django-cleanup==9.0.0
matplotlib
numpy